linted the project and add one custom hook
