+++ .eslintignore
@@ -1,0 +1,5 @@
+node_modules/
+dist/
+.prettierrc.js
+.eslintrc.js
+env.d.ts
