+++ index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/public/icon.jpeg" />
+    <link rel="icon" type="image/svg+xml" href="/icon.jpeg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Codebridge Test</title>
   </head>
