+++ package.json
@@ -6,6 +6,9 @@
   "scripts": {
     "dev": "vite --port 4000",
     "build": "tsc && vite build",
+    "lint": "eslint . --ext .ts,.tsx",
+    "pretty": "prettier --write ./src",
+    "format": "npm run lint && npm run pretty",
     "preview": "vite preview"
   },
   "dependencies": {
@@ -25,7 +28,16 @@
     "@types/qs": "^6.9.7",
     "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.9",
+    "@typescript-eslint/eslint-plugin": "^5.48.1",
+    "@typescript-eslint/parser": "^5.48.1",
     "@vitejs/plugin-react": "^3.0.0",
+    "eslint": "^8.31.0",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-plugin-import": "^2.27.4",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-react": "^7.32.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "prettier": "^2.8.3",
     "sass": "^1.57.1",
     "typescript": "^4.9.3",
     "vite": "^4.0.0"
