+++ src/hooks/useDebounce.ts
@@ -1,11 +1,11 @@
-import { useEffect } from 'react'
+import { useEffect, useRef } from 'react'
 
 const useDebounce = <T extends (...args: any[]) => void>(fn: T, dependency: unknown[], delay = 500) => {
-  let timer: NodeJS.Timer
+  const timer = useRef<NodeJS.Timer>()
   useEffect(() => {
-    clearTimeout(timer)
-    timer = setTimeout(() => fn(), delay)
-    return () => clearTimeout(timer)
+    clearTimeout(timer.current)
+    timer.current = setTimeout(() => fn(), delay)
+    return () => clearTimeout(timer.current)
   }, [dependency])
 }
